WIP try to separate status.py
